feat: interactive questions message type (#230)
* feat: add interactive questions message type

Add a first-class multi-turn question card message type so agents can ask
structured choice questions and the mobile client renders clickable
single/multi-select UI instead of plain text.

Agent emits via new MCP tool orchestrator__ask_questions; answers flow
back over Socket.IO (message.answers). On submission the backend
persists content.answers, broadcasts message.updated to flip the card
read-only, then synthesizes a user text message so the conversation
continues through the normal streaming pipeline.

Also flip all mobile-proto int64 counters/pagination/size fields to
int32 so proto3 JSON encodes them as numbers (not strings), fixing a
Dart type-cast crash in the mobile client.


* fix: address review findings on interactive questions

- Revert int32 over-downgrade: token counters, file size, and
  pagination total/offset now use double (JSON number, 53-bit precision)
  instead of int32. Eliminates silent truncation at 2.1B / 2GB while
  keeping the Dart client happy.
- Add input caps in AskQuestions: max 10 turns, 20 questions/turn,
  2KB prompt, 256B option label, 128B turn label. Prevents JSONB-bloat DoS.
- AllowCustom is now plumbed through from the MCP input instead of
  being hard-coded true; default false.
- RespondToQuestions: reject replay (already-answered check), require
  at least one selected option or custom text, enforce max custom text
  length (4KB), surface SendMessage follow-up failure via a distinct
  error code instead of silently swallowing.
- Stop leaking raw *merrors.Error.Error() over the socket; use
  PublicMessage mapping so DB/driver details don't reach clients.
- Typed Mode enum in service-layer params.
- Delete unused convert.QuestionAnswersToDomain + its tests.
- Error messages: lowercase, unpunctuated, no embedded quotes.
- Option ID base tied to alphabet via named const (maxOptions = 'Z'-'A'+1).


* refactor: address SonarQube criticals

- Extract validateAnswer + per-check helpers to cut validateQuestionAnswers
  cognitive complexity from 30 → <15.
- Extract formatAnswerLine to keep formatAnswersAsUserMessage under the
  complexity cap.
- Split MessageContentToProto into linear calls to actionsToProto,
  questionsToProto, questionAnswersToProto, argsToStructProto.
- Extract parseAnswerEntry + parseOptionIDs from parseMessageAnswersPayload.
- Deduplicate "question " literal via questionErrorf helper and const.
- Suppress the remaining int32(TurnIndex) cast with a dual #nosec/NOSONAR
  comment (turn count is capped at maxTurnsPerMessage).


---------

+++ internal/orchestrator/server/handler/usage.go
@@ -37,11 +37,11 @@ func UserUsageSummary(c *Context) http.HandlerFunc {
 
 		WriteProtoSuccess(w, http.StatusOK, &mobilev1.UserUsageSummaryResponse{
 			CurrentPeriod:        period,
-			TokensUsed:           tokensUsed,
-			PromptTokensUsed:     counters.PromptTokensUsed,
-			CompletionTokensUsed: counters.CompletionTokensUsed,
-			RequestCount:         int64(counters.RequestCount),
-			TokenLimit:           tokenLimit,
+			TokensUsed:           float64(tokensUsed),
+			PromptTokensUsed:     float64(counters.PromptTokensUsed),
+			CompletionTokensUsed: float64(counters.CompletionTokensUsed),
+			RequestCount:         float64(counters.RequestCount),
+			TokenLimit:           float64(tokenLimit),
 			PlanId:               counters.PlanID,
 		})
 	}
@@ -83,11 +83,11 @@ func WorkspaceUsage(c *Context) http.HandlerFunc {
 
 		WriteProtoSuccess(w, http.StatusOK, &mobilev1.WorkspaceUsageResponse{
 			Period:               period,
-			TokensUsed:           counters.TokensUsed,
-			PromptTokensUsed:     counters.PromptTokensUsed,
-			CompletionTokensUsed: counters.CompletionTokensUsed,
-			RequestCount:         int64(counters.RequestCount),
-			TokenLimit:           tokenLimit,
+			TokensUsed:           float64(counters.TokensUsed),
+			PromptTokensUsed:     float64(counters.PromptTokensUsed),
+			CompletionTokensUsed: float64(counters.CompletionTokensUsed),
+			RequestCount:         float64(counters.RequestCount),
+			TokenLimit:           float64(tokenLimit),
 		})
 	}
 }
